+++ api/teams/edit/modify-teams.php
@@ -1,19 +1,27 @@
 <?php
 
-function get_existing_team($id): array {
+function get_existing_team($id): null | array {
     $conn = db_connect();
+
+    if (!is_numeric($id)) {
+        return null;
+    }
 
     $result = $conn->query("SELECT * FROM teams WHERE tm_id = '$id'");
     $row = $result->fetch_assoc();
 
     return array(
         "name" => $row['tm_name'],
-        "hex" => $row["tm_hex_color"]
+        "hex" => '#' . $row["tm_hex_color"]
     );
 }
 
 function update_existing_team($id, $name, $hex): void {
     $conn = db_connect();
+
+    if (!is_numeric($id)) {
+        return;
+    }
 
     // Clean $hex so it is just the numeric hex code and no # in front
     $hex = str_replace("#", "", $hex);
@@ -47,6 +55,10 @@ function create_new_team($name, $hex): void {
 
 function delete_existing_team($id): void {
     $conn = db_connect();
+
+    if (!is_numeric($id)) {
+        return;
+    }
 
     // Delete all points added to the team
     $conn->query("DELETE FROM points WHERE pts_tm_id = '$id'");
